requirements.txt
amqp==5.0.6
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2
djangorestframework==3.12.4
fastapi==0.63.0
h11==0.12.0
html5lib==1.1
idna==2.10
kombu==5.0.2
lxml==4.6.3
prompt-toolkit==3.0.18
pydantic==1.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
